Вносит правки по замечаниям.
